# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alumni Tracking System API & Web Directory",
    "version": "1.1.0",
    "description": "Complete documentation for the Alumni Tracking System featuring both MVC Web Views (/users) and RESTful JSON APIs (/api/users), developed for the Web Programming course at Istanbul University."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Users (MVC Web Views)",
      "description": "Server-side rendered HTML views under /users"
    },
    {
      "name": "Users (REST API)",
      "description": "JSON API endpoints under /api/users"
    },
    {
      "name": "System",
      "description": "Health checks and system diagnostics"
    },
    {
      "name": "Introductory Routes",
      "description": "Basic Express introductory routes"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users (MVC Web Views)"],
        "summary": "User Listing View (HTML)",
        "description": "Renders the HTML table view listing all registered alumni and students. Supports search and role filtering.",
        "parameters": [
          { "name": "search", "in": "query", "description": "Search keyword", "schema": { "type": "string" } },
          { "name": "role", "in": "query", "description": "Filter by role", "schema": { "type": "string", "enum": ["STUDENT", "ALUMNI", "ADMIN"] } }
        ],
        "responses": {
          "200": { "description": "HTML web page with user list", "content": { "text/html": {} } }
        }
      },
      "post": {
        "tags": ["Users (MVC Web Views)"],
        "summary": "Create User via Web Form",
        "description": "Receives user creation form data, creates the user in-memory, and redirects to the user's profile view.",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "302": { "description": "Redirects to /users/{id} on success" },
          "400": { "description": "Re-renders form with validation error", "content": { "text/html": {} } }
        }
      }
    },
    "/users/new": {
      "get": {
        "tags": ["Users (MVC Web Views)"],
        "summary": "Render User Creation Form (HTML)",
        "description": "Renders the HTML form page to create a new user.",
        "responses": {
          "200": { "description": "HTML user creation form", "content": { "text/html": {} } }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users (MVC Web Views)"],
        "summary": "User Details View (HTML)",
        "description": "Renders the detailed profile view of a specific user.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "HTML user profile page", "content": { "text/html": {} } },
          "404": { "description": "HTML error page (User not found)", "content": { "text/html": {} } }
        }
      },
      "put": {
        "tags": ["Users (MVC Web Views)"],
        "summary": "Update User via Web Form",
        "description": "Handles full update from the edit form (supports _method=PUT or direct POST).",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "302": { "description": "Redirects to /users/{id} on success" },
          "400": { "description": "Re-renders edit form with validation error", "content": { "text/html": {} } }
        }
      },
      "delete": {
        "tags": ["Users (MVC Web Views)"],
        "summary": "Delete User via Web Form",
        "description": "Removes a user and redirects to the user directory /users.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "302": { "description": "Redirects to /users" }
        }
      }
    },
    "/users/{id}/edit": {
      "get": {
        "tags": ["Users (MVC Web Views)"],
        "summary": "Render User Edit Form (HTML)",
        "description": "Renders prefilled form to edit existing user details.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "HTML edit form", "content": { "text/html": {} } },
          "404": { "description": "HTML error page", "content": { "text/html": {} } }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Users (REST API)"],
        "summary": "List all users (JSON)",
        "description": "Returns array of all users in the in-memory store in JSON format.",
        "responses": {
          "200": {
            "description": "Array of user objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "count": { "type": "integer", "example": 2 },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users (REST API)"],
        "summary": "Create user (JSON / Form)",
        "description": "Creates a new user. Supports JSON, URL-Encoded, and Multipart formats.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserInput" } },
            "application/x-www-form-urlencoded": { "schema": { "$ref": "#/components/schemas/UserInput" } }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "Email conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users (REST API)"],
        "summary": "Get user by ID (JSON)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "User found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "put": {
        "tags": ["Users (REST API)"],
        "summary": "Full update user by ID (JSON)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserInput" } },
            "application/x-www-form-urlencoded": { "schema": { "$ref": "#/components/schemas/UserInput" } }
          }
        },
        "responses": {
          "200": { "description": "User updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "tags": ["Users (REST API)"],
        "summary": "Partial update user by ID (JSON)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserPatchInput" } },
            "application/x-www-form-urlencoded": { "schema": { "$ref": "#/components/schemas/UserPatchInput" } }
          }
        },
        "responses": {
          "200": { "description": "User patched", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Users (REST API)"],
        "summary": "Delete user by ID (JSON)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": ["System"],
        "summary": "Check system health",
        "description": "Returns system health, uptime, and system diagnostics in JSON format.",
        "responses": {
          "200": { "description": "Health status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HealthResponse" } } } }
        }
      }
    },
    "/": {
      "get": {
        "tags": ["Introductory Routes"],
        "summary": "Root home page",
        "description": "Returns HTML home page for browsers or JSON for API clients.",
        "responses": {
          "200": { "description": "Home page response" }
        }
      }
    },
    "/hello": {
      "get": {
        "tags": ["Introductory Routes"],
        "summary": "Hello World greeting",
        "responses": { "200": { "description": "Greeting text" } }
      }
    },
    "/hello/{name}": {
      "get": {
        "tags": ["Introductory Routes"],
        "summary": "Personalized greeting",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Personalized greeting" } }
      }
    },
    "/sum/{number1}/{number2}": {
      "get": {
        "tags": ["Introductory Routes"],
        "summary": "Sum of two numbers",
        "parameters": [
          { "name": "number1", "in": "path", "required": true, "schema": { "type": "number" } },
          { "name": "number2", "in": "path", "required": true, "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Calculated sum" } }
      }
    },
    "/about": {
      "get": {
        "tags": ["Introductory Routes"],
        "summary": "About project page",
        "responses": { "200": { "description": "About HTML page" } }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "email": { "type": "string", "format": "email", "example": "emre@example.com" },
          "first_name": { "type": "string", "example": "Emre" },
          "last_name": { "type": "string", "example": "Yılmaz" },
          "role": { "type": "string", "enum": ["STUDENT", "ALUMNI", "ADMIN"], "example": "ALUMNI" },
          "department": { "type": "string", "example": "Computer Engineering" },
          "graduation_year": { "type": "integer", "example": 2024 },
          "is_verified": { "type": "boolean", "example": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "UserInput": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "ahmet@example.com" },
          "first_name": { "type": "string", "example": "Ahmet" },
          "last_name": { "type": "string", "example": "Kaya" },
          "role": { "type": "string", "enum": ["STUDENT", "ALUMNI", "ADMIN"], "default": "ALUMNI" },
          "department": { "type": "string", "example": "Computer Engineering" },
          "graduation_year": { "type": "integer", "example": 2024 }
        }
      },
      "UserPatchInput": {
        "type": "object",
        "properties": {
          "first_name": { "type": "string", "example": "Ahmet Can" },
          "department": { "type": "string", "example": "Software Engineering" },
          "graduation_year": { "type": "integer", "example": 2025 }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/User" }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "UP" },
          "timestamp": { "type": "string", "format": "date-time" },
          "uptime": { "type": "object" },
          "system": { "type": "object" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Validation Error" },
          "message": { "type": "string", "example": "E-posta (email) alanı zorunludur." }
        }
      }
    }
  }
}